[JAC-CLOUD]: Refactors and enhancements
