+++ jac-cloud/jac_cloud/tests/simple_graph.jac
@@ -39,84 +39,6 @@ class User:BaseUser: {
     }
 }
 
-
-:walker:detach_node:can:enter_B {
-    return here del --> [-->];
-}
-
-:walker:create_nested_node:can:enter_root {
-    n = Nested(
-        val=0,
-        arr=[],
-        json={},
-        parent=Parent(
-            val=1,
-            arr=[1],
-            json={"a": 1},
-            child=Child(
-                val=2,
-                arr=[1, 2],
-                json={"a": 1, "b": 2},
-                enum_field = Enum.C
-            ),
-            enum_field = Enum.B
-        ),
-        enum_field = Enum.A
-    );
-    here ++> n;
-    return n;
-}
-
-:walker:update_nested_node:can:enter_root {
-    nested = [-->(`?Nested)][0];
-    nested.parent.child.json["c"] = 3;
-    nested.parent.child.arr.append(3);
-    nested.parent.child.val = 3;
-    nested.parent.child.enum_field = Enum.A;
-    nested.parent.json["b"] = 2;
-    nested.parent.arr.append(2);
-    nested.parent.val = 2;
-    nested.parent.enum_field = Enum.C;
-    nested.json["a"] = 1;
-    nested.arr.append(1);
-    nested.val = 1;
-    nested.enum_field = Enum.B;
-    return nested;
-}
-
-:walker:update_nested_node:can:enter_nested {
-    here.parent.child.json["c"] = 3;
-    here.parent.child.arr.append(3);
-    here.parent.child.val = 3;
-    here.parent.child.enum_field = Enum.A;
-    here.parent.json["b"] = 2;
-    here.parent.arr.append(2);
-    here.parent.val = 2;
-    here.parent.enum_field = Enum.C;
-    here.json["a"] = 1;
-    here.arr.append(1);
-    here.val = 1;
-    here.enum_field = Enum.B;
-    return here;
-}
-
-:walker:detach_nested_node:can:enter_root {
-    return here del--> [-->(`?Nested)];
-}
-
-:walker:visit_nested_node:can:enter_root {
-    nesteds = [-->(`?Nested)];
-    if nesteds {
-        return [-->(`?Nested)][0];
-    } else {
-        return nesteds;
-    }
-}
-
-:walker:visit_nested_node:can:enter_nested {
-    return here;
-}
-
 walker manual_create_nested_node {
     can enter_root with `root entry {
         n = Nested(
@@ -254,78 +176,5 @@ walker manual_delete_nested_edge {
     } else {
         import:py from jac_cloud.core.architype {BaseAnchor}
         Jac.disallow_root(here, BaseAnchor.ref(self.root_id));
-    }
-}
-
-walker custom_report {
-    can enter1 with `root entry {
-        report 1;
-        report 2;
-        report 3;
-
-        Jac.report({
-            "testing": 1
-        }, True);
-    }
-
-    class __specs__ {
-        has auth: bool = False;
-    }
-}
-
-walker Walker {
-    class __specs__ {
-        has private: bool = True;
-    }
-}
-
-node Node {
-    has val: str;
-
-    can entry1 with entry {
-        return f"{self.val}-2";
-    }
-
-    # with 'visit_sequence' (string class reference) is not yet supported
-    can entry3 with Walker entry {
-        return f"{self.val}-3";
-    }
-
-    # with 'visit_sequence' (string class reference) is not yet supported
-    can exit1 with Walker exit {
-        return f"{self.val}-4";
-    }
-
-    can exit2 with exit {
-        return f"{self.val}-5";
-    }
-}
-walker visit_sequence:Walker: {
-    can entry1 with entry {
-        return "walker entry";
-    }
-
-    can entry2 with `root entry {
-        here ++> Node(val = "a");
-        here ++> Node(val = "b");
-        here ++> Node(val = "c");
-        visit [-->];
-        return "walker enter to root";
-    }
-
-    can entry3 with Node entry {
-        return f"{here.val}-1";
-    }
-
-    can exit1 with Node exit {
-        return f"{here.val}-6";
-    }
-
-    can exit2 with exit {
-        return "walker exit";
-    }
-
-    class __specs__ {
-        has auth: bool = False;
     }
 }
